requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.29.2
botocore==1.32.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
Deprecated==1.2.14
Django==4.2.6
django-environ==0.11.2
djangorestframework==3.14.0
ecdsa==0.18.0
environ==1.0
idna==3.4
jmespath==1.0.1
joserfc==0.9.0
JPype1==1.4.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jwcrypto==1.5.0
packaging==23.2
psycopg2-binary
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2023.3.post1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.1
rsa==4.9
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.18
wrapt==1.16.0
